3.3 замітив назви як і просили Circle,Rectangle,Dog,Point.
