+++ src/Lesson6/Rectangle.java
@@ -2,12 +2,12 @@ package Lesson6;
 
 import java.util.Scanner;
 
-public class Exercise_06_01 {
+public class Rectangle {
     public static void main(String[] args) {
         //1) Створити клас прямокутник з полями довжина, ширина. Додати методи обчислення площі, периметру.
         System.out.println("Вважай що прямокутник створено");
         Scanner sc = new Scanner(System.in);
-        System.out.println("Програма сама обрахує площу + _ +");
+        System.out.println("Програма сама обрахує площу ");
         System.out.print("Введіть розмір одної із сторін прямокутника : ");
         int line1 = sc.nextInt();
         System.out.print("Введіть розмір другої сторони прямокутника : ");
