+++ src/Lesson6/TestDog.java
@@ -2,7 +2,7 @@ package Lesson6;
 
 public class TestDog {
     public static void main(String[] args) {
-        Dog_06_03 myDog = new Dog_06_03("Пес ", 5, 10, "Пітбуль");
+        Dog myDog = new Dog("Пес ", 5, 10, "Пітбуль");
         System.out.println(myDog.getName() + " Сосиска ");
         System.out.println(myDog.getBreed() + " порода");
         System.out.println(myDog.getAge() + " років, вік собаки");
